+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Created settings controller
 * Added reject call and send text message when receiving a call
 * Added setting to ignore group messages
+* Added connection with pairing code in chatwoot
 
 ### Fixed